backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0
python-dotenv>=1.0,<2.0
PyMuPDF>=1.24,<2.0
beautifulsoup4>=4.12,<5.0
httpx>=0.27,<1.0
chromadb>=0.5,<1.0
scikit-learn>=1.5,<2.0
numpy>=1.26,<3.0
